FIX-DEV: Moved the model for code helper to a route called code -- Chat route reserved for the purpose of chats -- Linked the route for code to code helper -- Navigation fixed- Now expands on choosing code helper -- Code controller now controls the code route -- Cleaned the chat controller
